+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Development Version:
 
 - Added ``FutureWarning`` to use ``mode='r'`` as default when opening files.
   By `Ryan Grout <https://github.com/groutr>`_.
+- Moved handling of ``_nc4_non_coord_`` to ``h5netcdf.BaseVariable``.
+  By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
 
 Version 0.11.0 (April 20, 2021):